requirements.txt
certifi
chardet
docutils
filestack-python
idna
Pygments
requests
trafaret
urllib3
Kivy
Kivy-Garden